Move createPromptText function to utils.js
