+++ src/components/forms.js
@@ -1,4 +1,4 @@
-import {setUpperCase, render} from './utils.js';
+import {setUpperCase, render, createPromptText} from './utils.js';
 import {createEventsCards} from './cards-of-events.js';
 
 export {
@@ -404,19 +404,6 @@ const removeNewEventForm = () => {
   tripEventsForm.remove();
   newEventBtn.toggleAttribute(`disabled`);
   createPromptText();
-};
-
-// create the prompt:
-const createPromptText = () => {
-  const events = document.getElementsByClassName(`trip-sort`)[0];
-  const promptText = document.getElementsByClassName(`prompt`)[0];
-  if (!events && !promptText) {
-    const prompt = document.createElement(`h2`);
-    prompt.classList.add(`prompt`);
-    prompt.textContent = `Click New Event to create your first point`;
-    const tripEvents = document.querySelector(`.trip-events`);
-    tripEvents.append(prompt);
-  }
 };
 
 // handlers:
